Implement LinkmodFiles
There is a TX module which writes frames to a directory (via the
PIGEON_FILES_TX environment variable), and a twin that *reads* frames
(via the PIGEON_FILE_RX environment variable). Frames written to the
file are base64 encoded and we write as many as we can.

There are also some minor cleanups for linkmod_console and new functions
to help with pushing and popping frames in pigeon_link.